refactor(captp): breakout makeRemoteKit from engine
